refactor(dashboard): extract Court dashboard auth sub-router (F-B-202 PR-2/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-202.md

Sprint 2 modularization PR-2 of 10. Stacked on PR #839 (F-B-202 PR-1
helpers + template_env extract). Moves the authentication surface
(login / logout / first-boot password setup / OIDC federation) into
a per-feature sub-router.

What moves:

  app/dashboard/auth_routes.py (new, 407 LOC, 7 routes):
    GET  /dashboard/login           login form (or /setup redirect on
                                    first boot)
    POST /dashboard/login           password login submit
    POST /dashboard/logout          session clear
    GET  /dashboard/setup           first-boot admin password page
    POST /dashboard/setup           first-boot password submit
                                    (consume_bootstrap_token_and_set_password)
    GET  /dashboard/oidc/start      OIDC authorization code flow init
    GET  /dashboard/oidc/callback   OIDC provider redirect handler

  app/dashboard/router.py (2930 → 2614 LOC, -316 LOC):
    - 7 routes removed; breadcrumb comments at the old positions.
    - MIN_ADMIN_PASSWORD_LENGTH constant moved with the routes.
    - Sub-router wired via router.include_router(_auth_routes.router)
      near the router declaration so URLs stay byte-identical
      (auth_routes.py uses relative paths /login /setup /oidc/* and
      inherits the parent /dashboard prefix).

Behavioural delta: zero. URLs, response bodies, template renders,
session cookie semantics all unchanged. Verified by running the
existing 202-test dashboard suite green + 25/25 smoke (1 pre-existing
deselect on test_dashboard_approvals_nav, expected).

Pattern mirrors mcp_proxy/dashboard/ which already has 8 sub-routers
threaded through the same include_router shape (memory
feedback_dashboard_template_envs_per_router).

Tests:
  - pytest tests/test_dashboard*.py -q — 202/203 (1 deselect)
  - pytest tests/ --ignore=tests/integration --ignore=tests/e2e
    --deselect <known> -q — 4209 passed, 9 skipped, 32 xfailed
    (3 xdist shared-state flakes pass isolated)
  - ./sandbox/demo.sh full + ./sandbox/smoke.sh full — 25/25
    (B4.7 timing flake required one re-run, expected)

Stack status (router.py): 67 → 60 routes inline. 6 more sub-router
extraction PRs to land before router.py becomes the ~100 LOC
aggregator the audit recommends.

Next PR-3 (F-B-202 modularization): extract admin_settings_routes.py
(section 4, change-password endpoint, ~ 90 LOC, 3 routes).
